requirements.txt
aioredis==1.3.1
appdirs==1.4.4
asgiref==3.2.10
async-timeout==3.0.1
attrs==20.2.0
autobahn==20.7.1
Automat==20.2.0
certifi==2020.4.5.1
cffi==1.14.3
Django==3.1.1
djangorestframework==3.11.0
djangorestframework-simplejwt==4.4.0
gunicorn==20.0.4
Pillow==7.1.1
pipenv==2020.6.2
psycopg2==2.8.5
PyJWT==1.7.1
requests==2.23.0
sqlparse==0.3.1
whitenoise==5.0.1
